+++ packages/components/outline-core-alert/src/config.ts
@@ -1,3 +1,16 @@
 /**
  * @file Shared configuration for the `outline-core-alert` component.
  */
+export const coreAlertStatusTypes = [
+  'info',
+  'warning',
+  'error',
+  'success',
+] as const;
+export type CoreAlertStatusType = (typeof coreAlertStatusTypes)[number];
+
+// This can be useful for testing.
+export interface OutlineCoreAlertInterface extends HTMLElement {
+  status: CoreAlertStatusType;
+  isInteractive: boolean;
+}