+++ packages/components/outline-core-alert/src/outline-core-alert.ts
@@ -1,10 +1,15 @@
-import { html, TemplateResult, CSSResultGroup } from 'lit';
-import { customElement } from 'lit/decorators.js';
-
-// Our base component, which all others extend.
+import { html, TemplateResult, CSSResultGroup, css } from 'lit';
+import { customElement, property } from 'lit/decorators.js';
+import { ifDefined } from 'lit/directives/if-defined.js';
 import { OutlineElement } from '@phase2/outline-core';
-
-import componentStyles from './outline-core-alert.css.lit';
+import {
+  CoreAlertStatusType,
+  OutlineCoreAlertInterface,
+} from '@phase2/outline-core-alert';
+import { AdoptedStyleSheets } from '@phase2/outline-adopted-stylesheets-controller';
+import componentStyles from './style/outline-core-alert.css.lit';
+import componentVars from './style/outline-core-alert.vars.css.lit';
+import globalStyles from './style/outline-core-alert.lightDom.css.lit';
 
 /** The element name, reused throughout the codebase */
 const componentName = 'outline-core-alert';
@@ -15,21 +20,76 @@ const componentName = 'outline-core-alert';
  *
  * @element outline-core-alert
  * @extends OutlineElement
- * @slot - The default slot for this element.
- *
- * @todo: Complete component.
- * @todo: Update `../docs/outline-core-alert.mdx` to accurately document the component.
- * @todo: Run `yarn analyze` to generate README.md.
- * @todo: Update `package.json` to remove the `private` flag.
- * @todo: Update `.changeset/config.json` to remove this component from the `ignore` list.
- * @todo: Create PR for updated and completed component.
+ * @slot header - The header in the alert.
+ * @slot icon-start - The icon to display at the start of the alert.
+ * @slot default - The alert contents.
+ * @slot icon-end - The icon to display at the end of the alert.
+ * @cssprop --outline-alert--info-background: The background color for the info alert.
+ * @cssprop --outline-alert--info-text: The text color for the info alert.
+ * @cssprop --outline-alert--info-border: The border color for the info alert.
+ * @cssprop --outline-alert--success-background: The background color for the success alert.
+ * @cssprop --outline-alert--success-text: The text color for the success alert.
+ * @cssprop --outline-alert--success-border: The border color for the success alert.
+ * @cssprop --outline-alert--warning-background: The background color for the warning alert.
+ * @cssprop --outline-alert--warning-text: The text color for the warning alert.
+ * @cssprop --outline-alert--warning-border: The border color for the warning alert.
+ * @cssprop --outline-alert--error-background: The background color for the error alert.
+ * @cssprop --outline-alert--error-text: The text color for the error alert.
+ * @cssprop --outline-alert--error-border: The border color for the error alert.
+ * @todo: Make the alert styling more flexible.
  */
 @customElement(componentName)
-export class OutlineCoreAlert extends OutlineElement {
+export class OutlineCoreAlert
+  extends OutlineElement
+  implements OutlineCoreAlertInterface
+{
   static styles: CSSResultGroup = [componentStyles];
+  adoptedStylesheets: AdoptedStyleSheets;
+  debug = false;
+
+  @property({ type: String, attribute: 'status' })
+  status: CoreAlertStatusType = 'info';
+
+  /**
+   * This is important context for screen readers.
+   */
+  @property({ type: Boolean, attribute: 'is-interactive' })
+  isInteractive = false;
+
+  /**
+   * @see `outline-core-link` documentation as to the purpose of this method.
+   */
+  connectedCallback() {
+    super.connectedCallback();
+    this.adoptedStylesheets = new AdoptedStyleSheets(css`
+      outline-core-alert {
+        ${componentVars}
+      }
+      ${globalStyles}
+    `);
+    this.addController(this.adoptedStylesheets);
+  }
 
   render(): TemplateResult {
-    return html` <slot></slot> `;
+    // The `body` wrapper is used to circumvent limitations with styling the `:host` directly, such as applying borders.
+    return html`
+      <div
+        class="alert-body"
+        role="${this.isInteractive ? 'alertdialog' : 'alert'}"
+        aria-labelledby=${ifDefined(this.isInteractive ? 'message' : undefined)}
+      >
+        <div class="alert-content">
+          <slot name="icon-start"></slot>
+          <div class="alert-header">
+            <slot name="header">${this.status}</slot>
+          </div>
+          <slot name="icon-end"></slot>
+          <div class="message" id="message">
+            <slot></slot>
+          </div>
+        </div>
+      </div>
+    `;
   }
 }