+++ packages/documentation/outline-docs/index.ts
@@ -1,0 +1,23 @@
+/**
+ * @file Exports for the `outline-docs` package.
+ *
+ * Any export here is then available via the following import,
+ * which should be available in the consuming application:
+ *
+ * @code
+ * import { OutlineCoreAlert } from '@phase2/outline-core-alert';
+ * @endcode
+ *
+ * Any of the above exports can be used in the consuming application
+ * based on the contents of this file, and creativity of what functionality
+ * you wish to expose to the consuming application(s).
+ */
+
+// export const OutlineDocsStorybookDefaultParams = {
+//   viewMode: 'docs',
+//   previewTabs: {
+//     canvas: {
+//       hidden: true,
+//     },
+//   },
+// };