+++ packages/outline-core/src/outline-element/utilities/localize.ts
@@ -1,5 +1,0 @@
-// Register English as the default/fallback language
-import '~/translations/en';
-
-// Export functions from the localize lib so we have one central place to import them from
-// export * from '@shoelace-style/localize';