+++ packages/documentation/outline-docs/CHANGELOG.md
@@ -1,5 +1,15 @@
 # @phase2/outline-docs
 
+## 0.0.20
+
+### Patch Changes
+
+- 3ce056db: Update and reorganization
+- b880f153: Cleaup and organization.
+- Updated dependencies [3ce056db]
+- Updated dependencies [b880f153]
+  - @phase2/outline-core@0.2.6
+
 ## 0.0.19
 
 ### Patch Changes