+++ packages/documentation/outline-storybook/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @phase2/outline-storybook
 
+## 0.1.3
+
+### Patch Changes
+
+- ee19425e: Update for core link best practices.
+- 3ce056db: Update and reorganization
+- b880f153: Cleaup and organization.
+
+## 0.1.2
+
+### Patch Changes
+
+- Updates for outline-core-button.
+
 ## 0.1.1
 
 ### Patch Changes