+++ packages/outline-templates/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @phase2/outline-templates
 
+## 0.0.50
+
+### Patch Changes
+
+- ee19425e: Update for core link best practices.
+- 3ce056db: Update and reorganization
+- b880f153: Cleaup and organization.
+
 ## 0.0.49
 
 ### Patch Changes